requirements.txt
Django>=3.0.0
django-cors-headers>=3.2.0
djangorestframework>=3.11.0
gunicorn>=20.0.4
django-allauth==0.41.0
django-extensions==3.0.3
django-rest-auth==0.9.5
djangorestframework-jwt==1.11.0
drf-yasg